1、修改除项目列表外（使用name），其他使用"项目列表"接口的页面，全部使用fullName字段 2、"合同详情"界面的"绩效详情"部分中的"发布总额"显示更改为"合同总额-制作费-组稿费"(自己计算)、 3、"工作申请"功能中，选择客户、项目时，没有新增功能（导航栏右按钮） 4、"拒绝创建客户/项目操作"功能添加后续界面（能选择或填写客户/项目）

+++ LanTuOA/Class/Customer/Model/ContractListModel.swift
@@ -60,6 +60,6 @@ struct ContractListData: HandyJSON {
     var otherParams: String?
     var paybackMoney: Float = 0
     var endTime: Int = 0
-    var makeMoney: Int = 0
+    var makeMoney: Float = 0
     var startTime: Int = 0
 }
